+++ tests/testthat/test-conf_handle_config_reference.R
@@ -1,0 +1,3 @@
+test_that("conf_handle_config_reference() works", {
+  warning("TODO-20191027-5: Write test for `conf_handle_config_reference()`")
+})