+++ tests/testthat/test-conf_has_config_reference.R
@@ -1,0 +1,3 @@
+test_that("conf_has_config_reference() works", {
+  warning("TODO-20191027-1: Write test for `conf_has_config_reference()`")
+})