+++ tests/testthat/test-conf_resolve_config_reference.R
@@ -1,0 +1,3 @@
+test_that("conf_resolve_config_reference() works", {
+  warning("TODO-20191027-3: Write test for `conf_resolve_config_reference()`")
+})